requirements.txt
asgiref==3.4.1
autopep8==1.5.7
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
decorator==5.0.9
Django==3.2.7
django-cors-headers==3.8.0
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
ipython==7.27.0
jedi==0.18.0
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.10.0
PyJWT==2.2.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.1.0
urllib3==1.26.6
wcwidth==0.2.5
